+++ lua/kulala/cmd/formatter.lua
@@ -13,14 +13,15 @@ local function capitalize(str)
   return str:lower():gsub("^%l", string.upper):gsub("%-%l", string.upper)
 end
 
-local function trim(str)
-  str = str or ""
-  str = str:gsub("^%s+", ""):gsub("%s+$", ""):gsub("[ \t]+", " ")
-  return str
+local function trim(str, collapse)
+  str = (str or ""):gsub("^%s+", ""):gsub("%s+$", "")
+  return collapse and str:gsub("[ \t]+", " ") or str
 end
 
-local function get_text(node)
-  return trim(ts.get_node_text(node, buf))
+local function get_text(node, collapse)
+  collapse = collapse == nil and true or collapse
+  local text = ts.get_node_text(node, buf)
+  return trim(text, collapse) or text
 end
 
 local function get_fields(node, names)
@@ -267,7 +268,7 @@ format_rules = {
   end,
 
   ["raw_body"] = function(node)
-    local body = get_text(node)
+    local body = get_text(node, false)
 
     local request = current_section().request
     request.body = #request.body > 0 and (request.body .. "\n\n" .. body) or body
@@ -276,7 +277,7 @@ format_rules = {
   end,
 
   ["multipart_form_data"] = function(node)
-    local body = get_text(node)
+    local body = get_text(node, false)
 
     local request = current_section().request
     request.body = #request.body > 0 and (request.body .. "\n\n" .. body) or body
@@ -285,7 +286,7 @@ format_rules = {
   end,
 
   ["xml_body"] = function(node)
-    local body = get_text(node)
+    local body = get_text(node, false)
 
     local formatted = Formatter.xml(body) or body
     formatted = formatted:gsub("\n*$", "")
@@ -297,7 +298,7 @@ format_rules = {
   end,
 
   ["json_body"] = function(node)
-    local json = get_text(node)
+    local json = get_text(node, false)
 
     if format_opts.quote_json_variables then
       local lcurly, rcurly = "X7BX7B", "X7DX7D"
@@ -324,7 +325,8 @@ format_rules = {
   end,
 
   ["graphql_data"] = function(node)
-    local body = get_text(node)
+    local body = get_text(node, false)
+
     local formatted = Formatter.graphql(body, { sort = format_opts.sort.json }) or body
     formatted = formatted:gsub("\n*$", "")
 
@@ -344,16 +346,16 @@ format_rules = {
   end,
 
   ["pre_request_script"] = function(node)
-    local script = get_text(node)
-    script = script:gsub("\n", "\n  "):gsub("\n  %%}", "\n%%}")
+    local script = get_text(node, false)
+    script = script:gsub("\n", "\n  "):gsub("\n%s+%%}", "\n%%}")
 
     table.insert(current_section().request.pre_request_script, script)
     return script
   end,
 
   ["res_handler_script"] = function(node)
-    local script = get_text(node)
-    script = script:gsub("\n", "\n  "):gsub("\n  %%}", "\n%%}")
+    local script = get_text(node, false)
+    script = script:gsub("\n", "\n  "):gsub("\n%s+%%}", "\n%%}")
 
     table.insert(current_section().request.res_handler_script, script)
     return script