+++ lua/kulala/config/defaults.lua
@@ -64,6 +64,9 @@ local M = {
     },
   },
 
+  -- format json response when redirecting to file
+  format_json_on_redirect = true,
+
   scripts = {
     -- Resolves "NODE_PATH" environment variable for node scripts. Defaults to the first "node_modules" directory found upwards from "script_file_dir".
     node_path_resolver = nil, ---@type fun(http_file_dir: string, script_file_dir: string, script_data: ScriptData): string|nil
@@ -115,6 +118,9 @@ local M = {
     show_request_summary = true,
     -- disable notifications of script output
     disable_script_print_output = false,
+
+    -- do not show responses over maximum size, in bytes
+    max_response_size = 32000,
 
     report = {
       -- possible values: true | false | "on_error"